+++ CHANGELOG.md
@@ -1,3 +1,25 @@
+0.13.5 / 2019-03-27
+===================
+- Fix `RequestQueue` throttling at high concurrency.
+
+0.13.4 / 2019-03-26
+===================
+- Fix bug in `addWebhook` invocation.
+
+0.13.3 / 2019-03-21
+===================
+- Fix `puppeteerPoolOptions` object not being used in `PuppeteerCrawler`.
+
+0.13.2 / 2019-03-21
+===================
+- Fix `REQUEST_QUEUE_HEAD_MAX_LIMIT` is not defined error.
+
+0.13.1 / 2019-03-21
+===================
+- `Snapshotter` now marks Apify Client overloaded on the basis of 2nd retry errors.
+- Added `Apify.addWebhook()` to invoke a webhook when an actor run ends.
+  Currently this only works on the Apify Platform and will print a warning when ran locally.
+
 0.13.0 / 2019-03-14
 ===================
 - **BREAKING CHANGE:** Added `puppeteerOperationTimeoutSecs` option to `PuppeteerPool`.